refactor: use component to register search
Refs: CO-1741
