+++ src/app.test.tsx
@@ -13,7 +13,7 @@ import App from './app';
 import * as addComponentsToShell from './app-utils/add-shell-components';
 import * as registerShellActions from './app-utils/register-shell-actions';
 import * as registerShellIntegrations from './app-utils/register-shell-integrations';
-import * as useSearchView from './app-utils/search-registration';
+import * as useSearchRegisterer from './app-utils/use-search-registerer';
 import { generateFolder } from './carbonio-ui-commons/test/mocks/folders/folders-generator';
 import {
 	createAPIInterceptor,
@@ -78,12 +78,16 @@ describe('App', () => {
 			registerShellIntegrations,
 			'registerShellIntegrations'
 		);
-		const useSearchViewSpy = jest.spyOn(useSearchView, 'useSearchView');
 		setupTest(<App />);
 		expect(addComponentsToShellSpy).toHaveBeenCalled();
 		expect(registerShellActionSpy).toHaveBeenCalled();
 		expect(registerShellIntegrationsSpy).toHaveBeenCalled();
-		expect(useSearchViewSpy).toHaveBeenCalled();
+	});
+
+	it('should register the search', () => {
+		const useSearchRegistererSpy = jest.spyOn(useSearchRegisterer, 'useSearchRegisterer');
+		setupTest(<App />);
+		expect(useSearchRegistererSpy).toHaveBeenCalled();
 	});
 
 	it('should add the backup search route when the backup search messages are present', () => {
